Refactor the classes of the NestedLoop Batch Iterators.
We will soon implemenent non-NestedLoop batch iterators, so:
 - Introduce a new JoinIterator abstract class to extend from.
 - Move the CrossJoin logic into a separate iterator: CrossJoinNLBatchIterator
 - Introduce a new JoinIterators utlity class which creates the appropriate Iterator.
 - Rename iterators by adding the NL (nested loop).
