+++ dex/src/main/java/io/crate/data/join/RightJoinNLBatchIterator.java
@@ -27,50 +27,47 @@ import io.crate.data.BatchIterator;
 import java.util.function.Predicate;
 
 /**
- * Combination of left + right join:
+ * Nested Loop + additional loop afterwards to emit any rows that had no matches
  *
  * <pre>
  *     for (leftRow in left) {
  *         for (rightRow in right) {
- *             match?
- *               onRow
- *         }
- *         if (noMatches) {
- *             onRow (right-side-null)
+ *             if matched {
+ *                 markPosition(pos)
+ *                 onRow
+ *             }
  *         }
  *     }
  *
  *     for (rightRow in right) {
- *         if (noMatches) {
- *              onRow (left-side-null)
+ *         if (noMatch(position)) {
+ *             onRow (left-side-null)
  *         }
  *     }
  * </pre>
  */
-class FullOuterJoinBatchIterator<L, R, C> extends NestedLoopBatchIterator<L, R, C> {
+class RightJoinNLBatchIterator<L, R, C> extends JoinBatchIterator<L, R, C> {
 
     private final LuceneLongBitSetWrapper matchedRows = new LuceneLongBitSetWrapper();
     private final Predicate<C> joinCondition;
 
     private boolean postNL = false;
-    private boolean hadMatch = false;
     private int position = -1;
 
-    FullOuterJoinBatchIterator(BatchIterator<L> left,
-                               BatchIterator<R> right,
-                               ElementCombiner<L, R, C> combiner,
-                               Predicate<C> joinCondition) {
+    RightJoinNLBatchIterator(BatchIterator<L> left,
+                             BatchIterator<R> right,
+                             ElementCombiner<L, R, C> combiner,
+                             Predicate<C> joinCondition) {
         super(left, right, combiner);
         this.joinCondition = joinCondition;
     }
 
     @Override
     public void moveToStart() {
-        postNL = false;
-        hadMatch = false;
-        position = -1;
-        activeIt = left;
         super.moveToStart();
+        activeIt = left;
+        postNL = false;
+        position = -1;
     }
 
     @Override
@@ -118,7 +115,6 @@ class FullOuterJoinBatchIterator<L, R, C> extends NestedLoopBatchIterator<L, R, 
         while (tryMoveRight()) {
             position++;
             if (joinCondition.test(combiner.currentElement())) {
-                hadMatch = true;
                 matchedRows.set(position);
                 return true;
             }
@@ -128,12 +124,6 @@ class FullOuterJoinBatchIterator<L, R, C> extends NestedLoopBatchIterator<L, R, 
         }
         position = -1;
         right.moveToStart();
-        if (hadMatch == false) {
-            combiner.nullRight();
-            activeIt = left;
-            return true;
-        }
-        hadMatch = false;
         return null;
     }
